Create Welcome popup
A new component holding a welcome popup has been created.
This popup is to be shown when the user first enters the webpage.
This commit also modifies tailwind's animation variants and the Page component.

As of now, the component holds a title, three points describing SocketChat and the author.
For now, it shows every time you enter the page, but in the future that will be optional.

The Page component has been updated to include this welcome popup.
An error with the margin of the main container has been fixed.
This error created an "empty" blank area on the top of the screen.
Changing the margin top class to a div with the same height before the content fixed this.

The tailwind config file has been modified to add animation variants.
'hover' and 'motion-safe' variants were already there, 'group-hover' was added.

The links of each section of the welcome page are yet to be worked on.