Add GPL-3.0 license file and notice
**THIS LICENSE APPLIES TO ANY PREVIOUS CODE OR RELEASE IN THIS CODEBASE**

The license for this project is now set to GNU General Public License v3.0.
The license file is created with the full license text, adapted to the project.
A notice is also added to all code files in the codebase (.vue, .js, .css & .html).

For any inquiries about the license change, please contact hipycas+socketchat@gmail.com