Short features text add GPL-3.0 badge to Welcome popup
The first modifiaction this commit makes is to reduce the features text length.
This is made because it was longer than the others (3 rather than 2).
This difference caused the popup to be scrollable on laptop screens, which was not intended.

As the GPL-3.0 license was added in the master branch, a badge has been added to the Welcome popup.
This SVG design was made by Christian Cadena and can be found at [the GNU webpage](https://www.gnu.org/graphics/license-logos.html)