Update README.md with project overview and installation instructions
Update the README.md to inculde three new sections: *Components of SocketChat*, *Software used* and *Self deploying*.
It is far from being complete, and it will be worked more on.

The *Components of SocketChat* section lays out what the repository contains.
It also lists which files are for the server and which for the client.

The *Software Used* lists the software/libraries that the project uses.
It also includes their library SPXD identifier and a reference to the NOTICE.md file.

The *Self deploying* section includes deep instructions in installing and running a SocketChat instance.
It includes a walk thorugh all steps, a TL;DR and some additional information.

The title of the yet to be worked on *Using SocketChat* section was added too.
An incomplete index was also added under *README index* section.