Add GPLv3 badge to Welcom popup and start work on Changes popup
In the Welcome popup, on the bottom-left, a GPLv3 badge/logo was added.
This is due to the decision of licensing the project under GPLv3.

Work has started on the Changes popup, which will show when the version changes.
This will be made functional by storing the last version used in local storage and checking against the server.
A CHANGELOG.md file has been also created, although it is not fully functional yet.